refactor(parser): create new package for parsers
